Refactor FileCache and remove FileCacheEx
