+++ cmake/vars.cmake
@@ -103,7 +103,7 @@ set(HTTP_SERVER_HEADERS
     http/server/HttpContext.h
     http/server/HttpResponseWriter.h
     http/server/WebSocketServer.h
-    http/server/FileCacheEx.h
+    http/server/FileCache.h
 )
 
 set(MQTT_HEADERS
